[fix] 폴더명 수정 및 header 적용
snake_case 로 폴더명 통일
내 목표 생성 화면 header 추가
