+++ sidebar/src/components/UI/atoms/input_text/InputText.js
@@ -17,7 +17,7 @@ const InputText = ({ style, variant, label, value, defaultValue, onChangeHandler
       defaultValue={defaultValue}
       onChange={onChangeHandler}
       InputProps={InputProps}
-      multiline="true"
+      multiline
     />
   );
 };
