+++ sidebar/src/components/UI/molecules/bucket_search_bar/BucketSearchBar.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import InputText from '../../atoms/InputText/InputText';
+import InputText from '../../atoms/input_text/InputText';
 
 const BucketSearchBar = ({ onChangeHandler }) => {
   const style = {
