+++ sidebar/src/components/UI/molecules/detail_adder/DetailAdder.js
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import InputAdornment from '@material-ui/core/InputAdornment';
 import AddCircleIcon from '@material-ui/icons/AddCircle';
 import { connect } from 'react-redux';
-import InputText from '../../atoms/InputText/InputText';
-import StyledButton from '../../atoms/StyledButton/StyledButton';
+import InputText from '../../atoms/input_text/InputText';
+import StyledButton from '../../atoms/styled_button/StyledButton';
 import { addDetailAction } from '../../../../modules/actions/createbucket';
 
 const DetailAdder = ({ addDetailActionConnect }) => {
