+++ sidebar/src/components/UI/molecules/reset_button/ResetButton.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import StyledButton from '../../atoms/StyledButton/StyledButton';
+import StyledButton from '../../atoms/styled_button/StyledButton';
 import { resetBucketAction } from '../../../../modules/actions/createbucket';
 
 const ResetButton = ({ resetBucketActionConnect }) => {
