+++ sidebar/src/components/UI/organisms/Buttons/Buttons.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import BucketCreateButton from '../../molecules/BucketCreateButton/BucketCreateButton';
-import ResetButton from '../../molecules/ResetButton/ResetButton';
+import BucketCreateButton from '../../molecules/bucket_create_button/BucketCreateButton';
+import ResetButton from '../../molecules/reset_button/ResetButton';
 
 const ButtonsWrapper = styled.div`
   display: flex;
