+++ sidebar/src/components/UI/organisms/bucket_contents/BucketContents.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import BucketInputText from '../../molecules/BucketInputText/BucketInputText';
+import BucketInputText from '../../molecules/bucket_input_text/BucketInputText';
 import { inputDescAction, inputTitleAction } from '../../../../modules/actions/createbucket';
 
 const BucketContentsWrapper = styled.div`
