+++ sidebar/src/components/UI/organisms/detail_list/DetailList.js
@@ -1,15 +1,16 @@
 import React from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import DetailAdder from '../../molecules/DetailAdder/DetailAdder';
-import DetailListItem from '../../molecules/DetailListItem/DetailListItem';
+import DetailAdder from '../../molecules/detail_adder/DetailAdder';
+import DetailListItem from '../../molecules/detail_list_item/DetailListItem';
 import { addDetailAction, removeDetailAction } from '../../../../modules/actions/createbucket';
+
+const DetailListWrapper = styled.div`
+  padding: 15px;
+`;
 
 const DetailList = ({ details }) => {
   // TODO : state로 바꿔야함
-  const DetailListWrapper = styled.div`
-    padding: 15px;
-  `;
   const Details = (items) => {
     return items.map((item) => {
       return <DetailListItem detail={item} />;
